statestrike/requirements.txt
openenv-core==0.1.0
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
httpx==0.27.0
pydantic==2.7.1
streamlit==1.35.0
plotly==5.22.0
python-dotenv==1.0.1
pytest==8.2.0
pytest-asyncio==0.23.7
rich==13.7.1
websockets==12.0
portalocker==2.8.2
